feat: rename core modules to minimal-repl and minimal-render for maximum safety
