+++ packages/opencode/src/cli/cmd/minimal-render.ts
